sync-github-releases: extract pure release logic into release-plan.ts
getReleasePlan and its pure helpers (getTagName, isOwnedTag, withSourceOfTruth,
chooseCreateCommitish, and the plan types) are the dependency-free domain logic
that decides what to create/update/delete and how each release is named, tagged
and described. Group them in one file so the pure, unit-tested core is separate
from the I/O orchestration in index.ts.
